+++ app/src/main/java/com/dede/android_eggs/views/main/compose/EasterEggGroupSelector.kt
@@ -12,12 +12,17 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
+import com.dede.android_eggs.composable.appbar.LocalHazeState
 import com.dede.android_eggs.views.main.util.EasterEggHelp
 import com.dede.basic.provider.BaseEasterEgg
 import com.dede.basic.provider.EasterEggGroup
+import dev.chrisbanes.haze.HazeInput
+import dev.chrisbanes.haze.blur.hazeBlur
+import dev.chrisbanes.haze.blur.materials.HazeMaterials
 
 
 @Composable
@@ -30,6 +35,7 @@ fun Modifier.withEasterEggGroupSelector(
     }
 
     val context = LocalContext.current
+    val hazeState = LocalHazeState.current
     var expanded by remember { mutableStateOf(false) }
     DropdownMenu(
         expanded = expanded,
@@ -37,7 +43,13 @@ fun Modifier.withEasterEggGroupSelector(
             expanded = false
         },
         shape = shapes.extraLarge,
-        offset = DpOffset(x = (-12).dp, y = 12.dp),
+        offset = DpOffset(x = (-12).dp, y = 10.dp),
+        containerColor = Color.Transparent,
+        modifier = Modifier.hazeBlur(
+            input = HazeInput.Sources(hazeState),
+            style = HazeMaterials.ultraThin(),
+        ),
+        shadowElevation = 6.dp,
     ) {
         base.eggs.forEachIndexed { index, egg ->
             val menuTitle = EasterEggHelp.VersionFormatter